irqchip/qcom-pdc: Add support for auxiliary GPIO interrupt controller
PDC v3.0+ optionally supports operation using an "auxiliary" GPIO interrupt
controller, where incoming GPIO interrupts are latched inside the PDC and
can be also masked/unmasked at the PDC level in addition to the GIC. This
feature can be enabled/disabled at runtime, but only the secure world can
change the configuration.

The mobile chipsets we have focused on so far have not made use of this
feature. However, on X1E80100 this feature is enabled by default and not
all firmware versions provide a way to disable it. To support all X1E80100
devices, we need to implement support for this feature in the Linux driver.

Implement the irqchip callbacks to enable this. We can detect which of the
PDC pins are GPIOs by trying to mask them using the additional register
bits. If the bit sticks, it's a GPIO and we need to unmask/acknowledge it
later.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>
---
TODO:

 - Fix or waive tlmm_test_rising_while_disabled failure. This test expects
   that GPIOs are still monitored when the IRQ is disabled. This doesn't
   easily work for the PDC. Keeping the IRQ enabled for monitoring would
   also trigger wakeup during suspend. The behavior of disabled IRQs does
   not seem to be clearly defined upstream and needs further discussion.

[   38.835973]     KTAP version 1
[   38.839149]     # Subtest: tlmm-test
[   38.842861]     # module: tlmm_test
[   38.842862]     1..17
[   38.954385]     ok 1 tlmm_test_silent_rising
[   39.064398]     ok 2 tlmm_test_silent_falling
[   39.178016]     ok 3 tlmm_test_silent_low
[   39.292140]     ok 4 tlmm_test_silent_high
[   39.783047]     ok 5 tlmm_test_rising
[   40.279180]     ok 6 tlmm_test_falling
[   40.635278]     ok 7 tlmm_test_high
[   40.993187]     ok 8 tlmm_test_low
[   41.131915]     ok 9 tlmm_test_rising_in_handler
[   41.273154]     ok 10 tlmm_test_falling_in_handler
[   41.767178]     ok 11 tlmm_test_thread_rising
[   42.267246]     ok 12 tlmm_test_thread_falling
[   42.525150]     ok 13 tlmm_test_thread_high
[   42.778360]     ok 14 tlmm_test_thread_low
[   42.927776]     ok 15 tlmm_test_thread_rising_in_handler
[   43.065504]     ok 16 tlmm_test_thread_falling_in_handler
[   43.166786]     # tlmm_test_rising_while_disabled: ASSERTION FAILED at drivers/pinctrl/qcom/tlmm-test.c:542
[   43.166786]     Expected atomic_read(&priv->intr_count) == 1, but
[   43.166786]         atomic_read(&priv->intr_count) == 0 (0x0)
[   43.195709]     not ok 17 tlmm_test_rising_while_disabled
[   43.195915] # tlmm-test: pass:16 fail:1 skip:0 total:17