pinctrl: qcom: Acknowledge IRQs for PDC auxiliary interrupt controller
When using the "auxiliary" GPIO interrupt controller feature in the PDC,
PDC needs to acknowledge incoming interrupts to clear the latched interrupt
status. For level-triggered IRQs this happens automatically in irq_eoi(),
but for edge-triggered IRQs this needs to happen as early as possible in
the IRQ handler.

Implement this by using handle_fasteoi_ack_irq() as IRQ handler in this
situation and forward the irq_ack() callback to the parent IRQ chip.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>