arm64: dts: qcom: x1e80100-crd: Add board-specific thermal zones
Add thermal zones for the 7 thermistors present on the CRD, which can be
read using the QCOM SPMI PMIC5 Gen3 ADC:

 - SYS_THERM0: SoC
 - SYS_THERM1: Keyboard hot spot (skin temperature)
 - SYS_THERM2: Back cover (skin temperature)
 - SYS_THERM3: Charging PMICs
 - SYS_THERM4: Modem
 - SYS_THERM5: Left side of the case (east)
 - SYS_THERM6: Right side of the case (west)

For now define only the critical trip points, as defined in the ACPI
tables.

NOTE: A critical temperature of 115°C does not really make sense for the
skin temperature sensors. It will cause potential skin burns long before
that. Those need more work, also for throttling if the fan cannot cool the
device enough to remain within safe bounds.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>